ass_2/Tiwana_Teg_Singh_Cloud_Assignment_2/LLM-CODEBASE-DOCUMENTOR-main/requirements.txt
pocketflow>=0.0.1
pyyaml>=6.0
requests>=2.28.0
gitpython>=3.1.0
google-cloud-aiplatform>=1.25.0
google-genai>=1.9.0
python-dotenv>=1.0.0
streamlit>=1.30.0
anthropic>=0.22.0
openai>=1.8.0
zipfile36>=0.1.3
